* 'PR/hexdump-color-overflow' of https://github.com/karelzak/util-linux-work:
  tests: (hexdump) use arrays for OPTS and ADDRFMT
  hexdump: fix buffer overflow in color_cond()

+++ text-utils/hexdump.h
@@ -39,8 +39,8 @@ struct hexdump_clr {
 	struct list_head colorlist;	/* next color unit */
 	const char *fmt;		/* the color, UL_COLOR_* */
 	off_t offt;			/* offset of range where unit is valid... */
-	int range;			/* ... and range length */
-	int val;			/* value ... */
+	size_t range;			/* ... and range length */
+	int64_t val;			/* value ... */
 	char *str;			/* ... or string to match */
 	int invert;			/* invert condition? */
 };
